update: move MemoryUtils into processor
